requirements.txt
# AI & Video
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Blockchain
web3>=6.0.0
python-dotenv>=1.0.0
py-solc-x>=2.0.0

# Alerts and storage
requests>=2.31.0
python-telegram-bot>=20.0

# Testing
pytest>=7.4.0
